+++ src/main/java/com/example/demo/controller/teacher/GroupController.java
@@ -1,4 +1,4 @@
-package com.example.demo.controller;
+package com.example.demo.controller.teacher;
 
 import com.example.demo.dto.GroupDto;
 import com.example.demo.dto.GroupWithCurrentSizeDto;
