+++ src/main/java/com/example/demo/controller/teacher/RelativeController.java
@@ -1,4 +1,4 @@
-package com.example.demo.controller;
+package com.example.demo.controller.teacher;
 
 import com.example.demo.dto.RelativeDto;
 import com.example.demo.repository.RelativeRepository;
